moved some methods out of availbookingscontroller and into booking class
